Move api/server/router/volume to daemon/server/router/volume
