+++ daemon/command/daemon.go
@@ -23,7 +23,6 @@ import (
 	buildbackend "github.com/docker/docker/api/server/backend/build"
 	"github.com/docker/docker/api/server/middleware"
 	"github.com/docker/docker/api/server/router"
-	"github.com/docker/docker/api/server/router/volume"
 	buildkit "github.com/docker/docker/builder/builder-next"
 	"github.com/docker/docker/builder/builder-next/exporter"
 	"github.com/docker/docker/builder/dockerfile"
@@ -45,6 +44,7 @@ import (
 	sessionrouter "github.com/docker/docker/daemon/server/router/session"
 	swarmrouter "github.com/docker/docker/daemon/server/router/swarm"
 	systemrouter "github.com/docker/docker/daemon/server/router/system"
+	"github.com/docker/docker/daemon/server/router/volume"
 	"github.com/docker/docker/dockerversion"
 	"github.com/docker/docker/internal/otelutil"
 	"github.com/docker/docker/libcontainerd/supervisor"
